feat: connect merkle root tracing with metrics
